CICD: prebuild docker image for test hardware
A large amount of time is spent in installing the toolchains directly
on the Raspberry Pi that does the hardware testing. Now we build a
Docker container for ARM which has already prebundled all those tools.
To do so, we use buildx together with a dummy `pio run` command; that is
necessary because `pio platform install` does not install optional
packages and does not offer any option to select or specify the correct
version for those.