Add wolfSSL support.
The library choice is controlled by `CONFIG_ESP_TLS_USE_WOLFSSL/MBEDTLS`, so it will automatically use whatever esp-idf uses.
It is possible to force it by setting `SPOOKY_USE_MBEDTLS/WOLFSSL`.

Additional checks ensure that wolfSSL is built with Des3 on. This requires a change to the bundled esp-wolfssl/port/user_settings.h, so we work around this by injecting our own with a higher priority using the `-iquote` compile option. This way we can keep esp-wolfssl as an untouched submodule.

The procedure for using wolfssl thus consists in
- cloning esp-wolfssl in the components/ subfolder of your project
- set `CONFIG_ESP_TLS_USE_WOLFSSL=y` in `sdkconfig.defaults`
- provide own override or modify  `user_settings.h` to *not* define `NO_DES3`

Furthermore, the number of DES contexts in mbedtls implementation has been reduced from 3 to 2, by observing that the only state they store is key and direction.