Implicit bool -> int conversion should be NUKED.
Blame is partially on me for using snake case and not camel case types, but essentially the bug is at test_desfire_files.cpp:110 (now 112). Instead of writing "free", the monostate constexpr instance of desfire::free_t, I typed free_access, which is a local bool variable.

This implicitly casts to int literal or whatever, but eventually constructs access_rights via the access_rights(std::uint8_t) constructor.

No warning, because it's ok in 2023 to convert a bool to an int. I might not be knowledgeable about this, but I think this should be just nuked away from the C++ standard and let everyone's code who relies on this simply break. Use a modern editor with clang-tidy on it or let your code die in oblivion.

So to prevent further confusions with free(void *), I renamed that to free_access and deleted all accidental bool conversion operators. This required to add a custom operator to cast to std::uint8_t, because that also is missing in 2023 from C++.