Fix auth and app workflow to match card behavior
I tested this on a card and it seems that the only
situation in which the active app is changed, is by
a select_application, otherwise it's persisted.
A failed select_application will bring us back to the
root app.
Independently, select_application logs out.
The only other command that logs out (aside from
authenticate) is change_key when operated on the
current active key.
Using the "change_key(prev, new)" overload does
not work when changing the same key, the other
overload must be used.